Refactor and optimize components; remove unused imports and code
- Removed "expo-image" from app.json dependencies.
- Cleaned up imports in grading and tabs layout files.
- Moved OrderItem component to its own file and optimized its implementation.
- Updated profile screen to use new TokenIcon and adjusted text styles.
- Enhanced walletHistory screen with new transaction icons and improved layout.
- Added new SVG icons for card send and receive actions.
