+++ app.json
@@ -64,7 +64,6 @@
           "isAccessMediaLocationEnabled": true
         }
       ],
-      "expo-image",
       "expo-web-browser",
       "expo-secure-store"
     ],
