+++ app/(grading)/_layout.tsx
@@ -1,5 +1,4 @@
 import { Stack } from "expo-router";
-import React from "react";
 
 export default function GradingLayout() {
   const commonHeaderOptions = {
