+++ app/(tabs)/orderHistory.tsx
@@ -18,119 +18,9 @@ import { Order, OrderStatus } from "@/types/order";
 import DateTimePicker, {
   DateTimePickerEvent,
 } from "@react-native-community/datetimepicker";
+import OrderItem from "@/components/organisms/order-item";
 
 // --- Component OrderItem ---
-type OrderItemProps = {
-  item: Order;
-  onPress: () => void;
-};
-const OrderItem = ({ item, onPress }: OrderItemProps) => {
-  const styleConfig = useMemo(() => {
-    switch (item.status) {
-      case "PAID":
-        return {
-          icon: "credit-card-check-outline" as const,
-          color: "text-green-500",
-          bgColor: "bg-green-100",
-        };
-      case "PENDING":
-        return {
-          icon: "clock-outline" as const,
-          color: "text-yellow-500",
-          bgColor: "bg-yellow-100",
-        };
-      case "CANCELLED":
-        return {
-          icon: "close-circle-outline" as const,
-          color: "text-red-500",
-          bgColor: "bg-red-100",
-        };
-      case "FAILED":
-        return {
-          icon: "alert-circle-outline" as const,
-          color: "text-red-500",
-          bgColor: "bg-red-100",
-        };
-      case "EXPIRED":
-        return {
-          icon: "clock-alert-outline" as const,
-          color: "text-gray-500",
-          bgColor: "bg-gray-100",
-        };
-      case "RETRY":
-        return {
-          icon: "refresh-circle" as const,
-          color: "text-orange-500",
-          bgColor: "bg-orange-100",
-        };
-      default:
-        return {
-          icon: "help-circle-outline" as const,
-          color: "text-gray-500",
-          bgColor: "bg-gray-100",
-        };
-    }
-  }, [item.status]);
-
-  const formattedDate = new Date(item.createdAt).toLocaleString("vi-VN", {
-    day: "2-digit",
-    month: "2-digit",
-    year: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-  });
-
-  const getStatusText = (status: OrderStatus) => {
-    switch (status) {
-      case "PENDING":
-        return "Chờ xử lý";
-      case "PAID":
-        return "Đã thanh toán";
-      case "CANCELLED":
-        return "Đã hủy";
-      case "FAILED":
-        return "Thất bại";
-      case "EXPIRED":
-        return "Đã hết hạn";
-      case "RETRY":
-        return "Thử lại";
-      default:
-        return status;
-    }
-  };
-
-  return (
-    <TouchableOpacity
-      onPress={onPress}
-      className="flex-row items-center p-4 bg-white border-b border-gray-100 active:bg-gray-50"
-    >
-      <View
-        className={`w-12 h-12 rounded-full items-center justify-center ${styleConfig.bgColor}`}
-      >
-        <MaterialCommunityIcons
-          name={styleConfig.icon}
-          size={28}
-          className={styleConfig.color}
-        />
-      </View>
-      <View className="flex-1 mx-4">
-        <Text
-          className="text-base font-semibold text-gray-800"
-          numberOfLines={2}
-        >
-          {item.description}
-        </Text>
-        <Text className="text-sm text-gray-500 mt-1">{formattedDate}</Text>
-        <Text className={`text-sm font-medium mt-1 ${styleConfig.color}`}>
-          {getStatusText(item.status)}
-        </Text>
-      </View>
-      <Text className="text-lg font-bold text-gray-800">
-        {item.amount.toLocaleString("vi-VN")} VNĐ
-      </Text>
-    </TouchableOpacity>
-  );
-};
 
 // --- Component FilterBar ---
 const orderStatusOptions = [
@@ -150,7 +40,7 @@ const FilterBar = ({
   selectedValue: string;
   onSelect: (value: string) => void;
 }) => (
-  <View className="flex-row items-center p-2 bg-white">
+  <View className="flex-row items-center py-5 px-2 bg-white">
     <ScrollView
       horizontal
       showsHorizontalScrollIndicator={false}
